feat: add crudOperations for Account

+++ TD-wallet/src/main/java/com/wallet/tdwallet/model/Transaction.java
@@ -13,7 +13,7 @@ import java.time.ZonedDateTime;
 public class Transaction {
     private int transactionId;
     private String transactionLabel;
-    private int transactionAmount;
+    private double transactionAmount;
     private String transactionType;
     private LocalDateTime transactionDate;
 }
